requirements.txt
torch>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pyyaml>=6.0
numpy>=1.24.0
python-multipart>=0.0.6
